requirements.txt
opencv-python==4.8.1.78
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.27
easyocr==1.7.1
transformers==4.36.2
openai-whisper
numpy==1.24.3
Pillow==10.1.0
ffmpeg-python==0.2.0
